+++ include/oneapi/mkl.hpp
@@ -22,7 +22,7 @@
 
 #include "oneapi/mkl/types.hpp"
 
-#include "oneapi/mkl/blas/blas.hpp"
+#include "oneapi/mkl/blas.hpp"
 #include "oneapi/mkl/rng.hpp"
 
 #endif //_ONEMKL_HPP_
