+++ scripts/generate_backend_api.py
@@ -42,7 +42,7 @@ The command below will generate:
 "onemkl_blas_mklgpu.hpp" header with declaration of all backend library APIs.
 API from backend library will be called from "oneapi::mkl::mklgpu::blas" namespace.
 
-{script}  include/oneapi/mkl/blas/blas.hpp include/oneapi/mkl/blas/detail/mklgpu/onemkl_blas_mklgpu.hpp oneapi::mkl::mklgpu::blas
+{script}  include/oneapi/mkl/blas.hpp include/oneapi/mkl/blas/detail/mklgpu/onemkl_blas_mklgpu.hpp oneapi::mkl::mklgpu::blas
 '''.format(script = argv[0]))
 
 if len(argv) < 3:
